Moving datasets.py => __init__.py
